portfolio/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
crispy-tailwind==0.5.0
Django==4.2.5
django-crispy-forms==2.0
django-parler==2.3
django-ranged-response==0.2.0
django-rosetta==0.9.9
django-simple-captcha==0.5.18
fontawesomefree==6.4.2
idna==3.4
load-dotenv==0.1.0
Pillow==10.0.1
polib==1.2.0
python-dotenv==1.0.0
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.0.5
whitenoise==6.5.0
xlwt==1.3.0
